Update utils.js
Unless I'm misunderstanding the purpose of the lazyLoadingFactor, it appears that in the ThumbnailsView, thumbnails currently on screen are not loaded as expected, while most off the screen are. Reversing the sense of the boolean return value of the isOnScreen function seems to cause these thumbnails to be loaded as expected.

This behavior is seen with the default value of the lazyLoadingFactor (1.5).