Restricts an annotation to the bounds of the tileSource image.
Previously, an annotation could be forced out of the image canvas,
this change only allows the annotation size change (delta) to be
modified if the drag point is within the bounds of the image.

This commit pulls in the logic that previously existed but was changed
in e584432e477ee529094d3cc88c81594ac546a5d6.

Fixes #1111