Adding ability to customize SanatizeHtml function
The sanatizeHtml function is too restrictive. This is leading to needed
tags and attributes being stripped from the Html used in annotation view
text. These changes expose the configuration for the allowed tags and
attributes so they can be modified as part of the configuration object
passed to Mirador on initialization.

[ticket: 1350]