annotations: Make search annotations prettier
This switches the visualization of search annotations from the standard
outline-only display to a translucent filled rectangle that is more in
line with how highlighted text is usually rendered on images in digital
library search engines.

For this, we introduce a new `isSearch` option to the
`CanvasAnnotationDisplay` class that enables this way of rendering.