Fix canvas label display in ViewerInfo.
The component would always get be passed an `undefined` prop for the
label, since the selector in the container used the `canvasIdx` to
select the label, but the selector expected the identifier.  This was
missed in #2851 when the canvas identifier was passed to components that
were missing it.
Additionally, this commit removes the `canvasLabel` prop from the
`WindowCanvasNavigationControls` container, since the label is never
used in the actual component. (#3258)